+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [0.6.0]
 ### Added
  - `richmenu link`
+ - `richmenu unlink`
 ### Changed
  - Re-arrange columns in `richmenu list` and add `name` column
+ - Update dependencies
 
 ## [0.5.0]
 ### Added